third_party/blink/web_tests/fast/forms/select-popup/popup-menu-crash-on-style-update.html

<!DOCTYPE html>
<body>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<script src="../../../resources/testdriver.js"></script>
<script src="../../../resources/testdriver-vendor.js"></script>
<script src="../resources/picker-common.js"></script>
<select><option>option</option></select>
<script>
var test = async_test('Do not crash in InternalPopupMenu::update');
var select = document.querySelector('select');
openPickerDeprecatedJsTest(select, function() {
    setTimeout(function() {
        select.setAttribute('style', 'font-size: larger;');
        select.offsetWidth;
        select.style.display = 'none';
        setTimeout(function() {
            test.step(function() { assert_true(true, 'Did not crash.'); });
            test.done();
        }, 0);
    }, 0);
}, function() {
    // Do nothing on Mac and Android.
    test.done();
});
</script>
</body>
